requirements.txt
torch>=2.0.0
torchvision>=0.15.0
matplotlib>=3.7.0
numpy>=1.24.0
umap-learn>=0.5.3
scikit-learn>=1.2.0
tqdm>=4.65.0
